+++ NEWS.md
@@ -4,7 +4,8 @@ CohortDiagnostics (Develop)
 Changes:
 
 1. Error handling: Use [checkmate](https://CRAN.R-project.org/package=checkmate) R-package to provide more informative error messages.
-2. Refactor runCohortDiagnostics, to seperate the management of cohorts that need execution into an independent function \code{manageCohortsForExecution} 
+2. Refactor runCohortDiagnostics: added new function by refactoring exisitng private functions. This new function
+get the JSON and parameterized OHDSI SQL for the cohorts for which diagnostics has been requested \code{getCohortsJsonAndSql} 
 
 CohortDiagnostics 1.0.0
 =======================